Split `blocks/Series.tsx` into two files
I only moved stuff around and fixed imports, that's all. I did that so
that the reviewer can just skip this commit. It is preparation for
making a general `VideoListBlock`.
